requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
prophet==1.1.4  # Requires pystan/cmdstanpy

# NLP for Voice Commands
spacy==3.7.2
# After install: python -m spacy download en_core_web_sm

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Logging and Monitoring
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Environment Management
python-dotenv==1.0.0

# Optional: For advanced ML/Deep Learning
torch==2.1.1
transformers==4.35.2

# Time Series
statsmodels==0.14.0

# Caching
redis==5.0.1

# API Documentation
swagger-ui-bundle==0.0.9
